refactor: 빈 등록 설정을 위한 클래스의 패키지를 이동
빈을 등록하는데 사용하는 클래스를 `bean` 패키지 아래로 이동함.
요청 경로에 따른 매핑을 담당하는 클래스를 `web` 패키지 아래로 이동함.
