+++ src/main/java/codesquad/bean/Bean.java
@@ -1,4 +1,4 @@
-package codesquad.config;
+package codesquad.bean;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
