+++ src/main/java/codesquad/handler/MainHandler.java
@@ -7,6 +7,8 @@ import codesquad.message.HttpRequest;
 import codesquad.message.HttpStatusCode;
 import codesquad.model.User;
 import codesquad.util.ResourceUtils;
+import codesquad.web.ModelAndView;
+import codesquad.web.RequestMapping;
 import java.util.NoSuchElementException;
 
 public class MainHandler {
