+++ src/main/java/codesquad/view/Expression.java
@@ -1,6 +1,6 @@
 package codesquad.view;
 
-import codesquad.handler.ModelAndView;
+import codesquad.web.ModelAndView;
 
 public interface Expression {
     String handle(String template, ModelAndView mav);
