+++ src/main/java/codesquad/web/Handler.java
@@ -1,4 +1,4 @@
-package codesquad.handler;
+package codesquad.web;
 
 import codesquad.message.HttpRequest;
