+++ src/main/java/codesquad/web/StaticResourceHandler.java
@@ -1,4 +1,4 @@
-package codesquad.handler;
+package codesquad.web;
 
 import codesquad.message.HttpRequest;
 import codesquad.util.ResourceUtils;
