+++ src/test/java/codesquad/handler/LoginHandlerTest.java
@@ -9,6 +9,7 @@ import codesquad.message.HttpMethod;
 import codesquad.message.HttpRequest;
 import codesquad.message.HttpStatusCode;
 import codesquad.model.User;
+import codesquad.web.ModelAndView;
 import java.io.BufferedReader;
 import java.io.StringReader;
 import org.junit.jupiter.api.BeforeEach;
